# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flexi Route Logistics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/password:
    patch:
      operationId: changePassword
      tags: [auth]
      summary: Change current user password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChange"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- USERS ---
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List all drivers (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: isAvailable
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"

    post:
      operationId: createDriver
      tags: [drivers]
      summary: Create driver account (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Driver created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get driver by ID (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

    patch:
      operationId: updateDriver
      tags: [drivers]
      summary: Update driver (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

    delete:
      operationId: deleteDriver
      tags: [drivers]
      summary: Delete driver (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # --- DRIVER SELF-SERVICE ---
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get notifications for current user
      security:
        - bearerAuth: []
      parameters:
        - name: isRead
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # --- ADMIN ANALYTICS ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # --- AUTH SCHEMAS ---
    UserRegistration:
      type: object
      required: [name, email, password, phone]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        phone:
          type: string
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    PasswordChange:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8

    # --- USER SCHEMAS ---
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, driver, admin]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Driver:
      type: object
      required: [id, userId, isAvailable, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
        licenseNumber:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        totalDeliveries:
          type: integer
        createdAt:
          type: string
          format: date-time
    DriverInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        phone:
          type: string
        vehicleType:
          type: string
        licenseNumber:
          type: string
    DriverUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        vehicleType:
          type: string
        licenseNumber:
          type: string
        isAvailable:
          type: boolean

    # --- NOTIFICATION SCHEMAS ---
    Notification:
      type: object
      required: [id, userId, title, message, isRead, type, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        type:
          type: string
          enum: [shipment_update, payment_update, system, driver_assignment]
        relatedId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time

    # --- ADMIN ANALYTICS ---
...